product_service/requirements.txt
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.32.0
python-dotenv==1.1.1
SQLAlchemy[asyncio]==2.0.43
asyncpg==0.30.0